refactor: move appbar utils to separate folders
