+++ src/features/AppBar/Icons/Icons.styled.ts
@@ -1,0 +1,7 @@
+import styled from 'styled-components'
+import { spacings } from '../../../tokens/spacings'
+
+export const Icons = styled.div`
+  display: flex;
+  gap: ${spacings.SMALL};
+`
